Create .gitignore
I'm not sure if we should track the Pipfile.lock with Git, due to the whole cross-platform dependency issue. I will try to manually check if any of the current libraries could cause an issue.

This was created using gitignore.io, let me know if I should add anything else and generate a new file.